Proposal for patching Sorald rule 2164

+++ testcases/MathOnFloat.java
@@ -1,3 +1,4 @@
+
 // Test for rule s2164
 
 class MathOnFloat {
@@ -6,7 +7,7 @@ class MathOnFloat {
     void myMethod() {
         float a = 16777216.0f;
         float b = 1.0f;
-        float c = a + b; // Noncompliant {{Use a "double" or "BigDecimal" instead.}} yields 1.6777216E7 not 1.6777217E7
+        float c = (double) a + (double) b; // Noncompliant {{Use a "double" or "BigDecimal" instead.}} yields 1.6777216E7 not 1.6777217E7
 
         double d1 = a + b; // Noncompliant ; addition is still between 2 floats
         double d2 = a - b; // Noncompliant
